services/track-api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
PyJWT==2.10.1
redis==5.2.0
prometheus-client==0.21.1
